+++ wh_application/module/WebHemi2/Module.php
@@ -106,7 +106,7 @@ class Module implements
         // BEFORE the controller/action is being called we inject the correct layout
         $eventManager->attach(Mvc\MvcEvent::EVENT_DISPATCH,       ['WebHemi2\Event\LayoutEvent', 'preDispatch'],        10);
         // AFTER the controller/action is being called and have error we overwrite the default error pages
-        $eventManager->attach(Mvc\MvcEvent::EVENT_DISPATCH_ERROR, ['WebHemi2\Event\ErrorEvent',  'postDispatchError'], -50);
+        $eventManager->attach(Mvc\MvcEvent::EVENT_DISPATCH_ERROR, ['WebHemi2\Event\ErrorEvent',  'postDispatchError'], -150);
         // BEFORE rendering the output we change it, if it is an Ajax request
         $eventManager->attach(Mvc\MvcEvent::EVENT_RENDER,         ['WebHemi2\Event\AjaxEvent',   'preRender'],         -10);