+++ wh_application/module/WebHemi2/config/Admin.module.config.php
@@ -40,6 +40,12 @@ return [
         ],
         'template_map' => [
             'layout/layout' => __DIR__ . '/../resources/default/view/layout/admin.phtml',
+            'error/500'     => __DIR__ . '/../resources/default/view/error/500.phtml',
+            'error/403'     => __DIR__ . '/../resources/default/view/error/403.phtml',
+            'error/404'     => __DIR__ . '/../resources/default/view/error/404.phtml',
+            'block/header'  => __DIR__ . '/../resources/default/view/block/admin/header.phtml',
+            'block/footer'  => __DIR__ . '/../resources/default/view/block/admin/footer.phtml',
+            'block/menu'    => __DIR__ . '/../resources/default/view/block/admin/menu.phtml',
         ],
     ],
     'router' => [