+++ wh_application/module/WebHemi2/resources/default/static/js/login.js
@@ -74,7 +74,7 @@ $(document).ready(function() {
             url: '/login/',
             type: 'post',
             success: function(data) {
-                if (data.success) {.
+                if (data.success) {
                     location.href = 'http://' + DOMAIN + '/';
                 } else {
                     // remove all previous errors