+++ wh_application/module/WebHemi2/src/WebHemi2/Event/ErrorEvent.php
@@ -42,12 +42,6 @@ use Zend\Http\Response;
  */
 class ErrorEvent
 {
-    /** @var string $template */
-    public static $template = [
-        403 => 'error/403',
-        404 => 'error/404'
-    ];
-
     /**
      * Prepares the ACL error page
      *
@@ -66,19 +60,33 @@ class ErrorEvent
         switch ($error) {
             case 'error-unauthorized-controller':
             case 'error-unauthorized-route':
-                self::get403($event);
+                $layout = self::get403($event);
+                $code = 403;
                 break;
 
             default:
-                self::get404($event);
+                $code = 404;
+                $layout = self::get404($event);
         }
+
+        $event->setViewModel($layout);
+
+        /** @var Response $response */
+        $response = $event->getResponse();
+
+        // if no response object present, we create one
+        if (!$response) {
+            $response = new \HttpResponse();
+            $event->setResponse($response);
+        }
+        $response->setStatusCode($code);
     }
 
     /**
      * Prepares the 403 error page
      *
      * @param MvcEvent $event
-     * @return void
+     * @return ViewModel
      */
     protected static function get403(MvcEvent $event)
     {
@@ -103,81 +111,31 @@ class ErrorEvent
         $layout = $event->getViewModel();
         $layout->setVariable('title', '403 Forbidden');
 
-        if (ADMIN_MODULE == APPLICATION_MODULE
-            && $event->getApplication()->getServiceManager()->get('auth')->hasIdentity()
-        ) {
-            $headerBlock = new ViewModel();
-            $headerBlock->setTemplate('block/AdminHeaderBlock');
-
-            $menuBlock = new ViewModel();
-            $menuBlock->setTemplate('block/AdminMenuBlock');
-
-            $footerBlock = new ViewModel();
-            $footerBlock->setTemplate('block/AdminFooterBlock');
-
-            $layout->addChild($headerBlock, 'HeaderBlock')
-                ->addChild($menuBlock, 'MenuBlock')
-                ->addChild($footerBlock, 'FooterBlock');
-        }
-
         $model = new ViewModel($viewVariables);
-        $model->setTemplate(self::$template[403]);
+        $model->setTemplate('error/403');
         $model->setVariable('error', $error);
-        $layout->addChild($model);
+        $layout->addChild($model, 'content');
 
-        /** @var Response $response */
-        $response = $event->getResponse();
-
-        // if no response object present, we create one
-        if (!$response) {
-            $response = new \HttpResponse();
-            $event->setResponse($response);
-        }
-        $response->setStatusCode(403);
+        return $layout;
     }
 
     /**
      * Prepares the 404 error page
      *
      * @param MvcEvent $event
-     * @return void
+     * @return ViewModel
      */
     protected static function get404(MvcEvent $event)
     {
         // add our error page to the view model
         $layout = $event->getViewModel();
-        $layout->getVariable('title', '404 Not Found');
-
-        if (ADMIN_MODULE == APPLICATION_MODULE
-            && $event->getApplication()->getServiceManager()->get('auth')->hasIdentity()
-        ) {
-            $headerBlock = new ViewModel();
-            $headerBlock->setTemplate('block/AdminHeaderBlock');
-
-            $menuBlock = new ViewModel();
-            $menuBlock->setTemplate('block/AdminMenuBlock');
-
-            $footerBlock = new ViewModel();
-            $footerBlock->setTemplate('block/AdminFooterBlock');
-
-            $layout->addChild($headerBlock, 'HeaderBlock')
-                ->addChild($menuBlock, 'MenuBlock')
-                ->addChild($footerBlock, 'FooterBlock');
-        }
+        $layout->setVariable('title', '404 Not Found');
 
         $model = new ViewModel();
-        $model->setTemplate(self::$template[404]);
+        $model->setTemplate('error/404');
         $model->setVariable('reason', $event->getError());
         $layout->addChild($model);
 
-        /** @var Response $response */
-        $response = $event->getResponse();
-
-        // if no response object present, we create one
-        if (!$response) {
-            $response = new \HttpResponse();
-            $event->setResponse($response);
-        }
-        $response->setStatusCode(404);
+        return $layout;
     }
 }