+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 Change log
 ==========
 
+
+Version 2.0.3-p3
 ----------------
+- Add universal Ajax response generator 
+- Make admin login Ajax-based
+- Standardize view block and error page rendering
+- TODO fix issue with menus on error pages in the admin module
+- Tagging standalone version commits
+
 Version 2.0.3-p2
 ----------------
 - Added pre-commit git hook script